requirements.txt
beautifulsoup4
requests
yagmail
python-dotenv